pyproject.toml
[project]
name = "atmosphere"
version = "0.1.0"
description = "Real-time Bluesky firehose analytics with Spark, Iceberg, and Grafana"
requires-python = ">=3.11"
dependencies = [
    "pyspark>=4.0.0",
    "boto3>=1.35.0",
    "requests>=2.32.0",
    "typer>=0.12.0",
    "textual>=0.80.0",
    "python-dotenv>=1.0.0",
    "duckdb>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"
